refactor[frontend]: useValidationErrors to useFormatValidationErrors
